Split fixtures in Page module
