+++ saleor/page/tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .page import *  # noqa F403
+from .page_type import *  # noqa F403
+from .page_translation import *  # noqa F403
